~Moved XKCD from its own COG to the fun COG, as it fits better.
