+++ cogs/fun.py
@@ -2,11 +2,17 @@ from discord.ext import commands
 import discord
 import json
 import random
+import aiohttp
 
 
 class Fun(commands.Cog, name="Fun Commands"):
     def __init__(self, bot):
         self.bot = bot
+
+
+    ###################################
+    ### Skidy, Ashu, Dark's quotes. ###
+    ###################################
 
     @commands.command()
     async def skidy(self, ctx):
@@ -29,6 +35,10 @@ class Fun(commands.Cog, name="Fun Commands"):
         response.set_author(name="DarkStar7471",icon_url="https://i.imgur.com/jZ908d1.png")
         await ctx.send(embed=response)
 
+
+    #####################
+    ### HONK and BOOP ###
+    #####################
 
     @commands.command()
     async def honk(self, ctx):
@@ -56,6 +66,28 @@ class Fun(commands.Cog, name="Fun Commands"):
         else:
             return
 
+
+    ############
+    ### XKCD ###
+    ############
+
+    @commands.command()
+    async def xkcd(self,ctx):
+        comic_no = random.randint(1,1900)
+        url = f"http://xkcd.com/{comic_no}/info.0.json"
+        async with aiohttp.ClientSession() as session:
+            async with session.get(url) as data:
+                new_data = await data.read()
+                json_data = json.loads(new_data)
+                img = json_data.get("img")
+                title = json_data.get("title")
+                alt = json_data.get("alt")
+                response = discord.Embed(color=0xffb6b9)
+                response.add_field(name=title,value=alt)
+                response.set_image(url=img)
+                response.set_author(name="TryHackMe",icon_url="https://tryhackme.com/img/THMlogo.png")
+                response.set_footer(text="From the XKCD Official API!")
+        await ctx.send(embed=response)
