+++ config/channels.json
@@ -1,4 +1,4 @@
 {
 	"welcome":649001546985766946,
-	"announcements":546650767495397376
+	"announcements":649001546985766946
 }
