+++ config/dark.json
@@ -2,7 +2,7 @@
 	"quotes":[
 	"Listen here, you little ####.",
 	"*ahem* Can help you?",
-	"*sigh* What do you want.",
+	"*sigh* What do you want?",
 	"aight imma head out play Tarkov",
 	"Peace, dudes.",
 	"Very carefully, next question.",
