+++ config/guild.json
@@ -1,1 +1,1 @@
-{ "server":521382216299839518 }
+{ "server":647843704132075541 }
