+++ config/roles.json
@@ -1,22 +1,22 @@
 {
 	"ranks":[
-		638047504822173726, 
-		638047659541790760, 
-		638047662893039634, 
-		638047665766137876, 
-		638047668597162006, 
-		638047670878863390, 
-		638047673173409811, 
-		638047675845181461, 
-		638047678432804925, 
-		638047682232975371, 
-		638047685340954634, 
-		638047689380069404, 
-		638047691800313866
+		647881093579800576, 
+		647881098667491348, 
+		647881099334254611, 
+		647881100940541983, 
+		647881101477543944, 
+		647881101951369216, 
+		647921841293361173, 
+		647921846179594242, 
+		647921848977457163, 
+		647921850902511629, 
+		647921855507857448, 
+		647921857827176489, 
+		647921859899162625
 	],
 
-	"sub":538509395605061653,
+	"sub":647881944075468821,
 
-	"devLead":578349526574694421,
-	"dev":578056412790390785
+	"devLead":648977588462682128,
+	"dev":648977589784018983
 }
