+++ config/storage.json
@@ -1,1 +1,1 @@
-[{title:}]
+[{"image": "https://tryhackme-images.s3.amazonaws.com/room-icons/526fb97c8ede3330397e5cee20a8db6a.png", "title": "Avengers Blog", "description": "Hack into Tony Stark's machine! You will enumerate the machine, bypass a login portal via SQL injection and gain root access by command injection.", "code": "avengers", "users": 8, "tags": ["sqli", "rce", "avengers", "web", "linux"], "type": "walkthrough", "difficulty": "easy", "oneTimeFee": 0, "completedUsers": ["MrSeth6797", "PingTrip"], "upVotes": 3, "created": "2019-11-28T19:26:06.371Z"}, {"image": "https://tryhackme-images.s3.amazonaws.com/room-icons/efbb70493ba66dfbac4302c02ad8facf.jpeg", "title": "LazyAdmin", "description": "Easy linux machine to practice your skills", "code": "lazyadmin", "users": 14, "tags": ["security", "linux"], "type": "challenge", "difficulty": "easy", "oneTimeFee": 0, "completedUsers": ["haruue", "joona", "NinjaJc01", "PingTrip"], "upVotes": 6, "created": "2019-11-28T15:52:54.112Z"}, {"image": "https://tryhackme-images.s3.amazonaws.com/room-icons/ae431fbbd58e7f2fe55ea2f84f810801.png", "title": "Linux Privesc Playground", "description": "A linux privilege escalation playground, filled with badly configured superuser permissions, incorrectly set SUID bits, and more!", "code": "privescplayground", "users": 25, "tags": ["security", "privesc", "linux", "linenum"], "type": "challenge", "difficulty": "easy", "oneTimeFee": 0, "completedUsers": ["haruue", "pcmhdhp", "numl3er47", "virdoex", "NinjaJc01", "m1rf4k", "PingTrip", "nbyte", "nanoshi", "nanoshi", "t0thkr1s", "CirenX"], "upVotes": 5, "created": "2019-11-27T21:09:27.813Z"}, {"image": "https://tryhackme-images.s3.amazonaws.com/room-icons/90243c256ba1d4eb403bc206e376f238.jpeg", "title": "Blueprint", "description": "Hack into this Windows machine and escalate your privileges to Administrator.", "code": "blueprint", "users": 20, "tags": ["security", "windows"], "type": "challenge", "difficulty": "easy", "oneTimeFee": 0, "completedUsers": ["haruue"], "upVotes": 3, "created": "2019-11-27T20:06:55.445Z"}, {"image": "https://tryhackme-images.s3.amazonaws.com/room-icons/e5e8aa17c07012d2adc33f08d594ea10.jpeg", "title": "Basic Malware RE", "description": "This room aims towards helping everyone learn about the basics of \"Malware Reverse Engineering\".", "code": "basicmalwarere", "users": 21, "tags": ["security", "reverse engineering ", "malware", "malware analysis"], "type": "challenge", "difficulty": "medium", "oneTimeFee": 0, "completedUsers": ["w4tchd0g", "PingTrip", "t0thkr1s", "mrpvr"], "upVotes": 4, "created": "2019-11-26T16:09:17.360Z"}, {"image": "https://tryhackme-images.s3.amazonaws.com/room-icons/09e72853db5f1a9dd0bff63325acd99e.png", "title": "Reversing ELF", "description": "Room for beginner Reverse Engineering CTF players", "code": "reverselfiles", "users": 25, "tags": ["reverse engineering", "elf", "ctf", "challenge"], "type": "challenge", "difficulty": "easy", "oneTimeFee": 0, "completedUsers": ["PingTrip", "fsharp", "zilikos"], "upVotes": 3, "created": "2019-11-25T08:52:02.759Z"}]
